kubeproxy: implement kubeproxy installation
This changes implements the basic operator functionality that can
bootstrap a new cluster with kubeproxy. The operator currently
runs external to the cluster (on your local machine). For an example,
see the ./create-cluster.sh script that uses kinder.

The change adds 'ClusterCIDR' to the KubeProxy object. It would be
interesting to see if we can detect this value when running within the
cluster.